+++ NEWS.md
@@ -1,6 +1,8 @@
 # multid 0.1.0.9000
 
 * Development version
+* Fixed bug: renaming output in D_regularized_fold functions
+* Fixed joining data frames by fold.var in D_regularized_fold functions
 
 # multid 0.1.0