+++ NEWS.md
@@ -3,6 +3,7 @@
 * Development version
 * Fixed bug: renaming output in D_regularized_fold functions
 * Fixed joining data frames by fold.var in D_regularized_fold functions
+* Added statistical inference to d_pooled_sd
 
 # multid 0.1.0