+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fixed bug: renaming output in D_regularized_fold functions
 * Fixed joining data frames by fold.var in D_regularized_fold functions
 * Added statistical inference to d_pooled_sd
+* Added probabilities of correct classification option to D_regularized_out and D_regularized_fold_out
 
 # multid 0.1.0