+++ NEWS.md
@@ -1,11 +1,12 @@
-# multid 0.1.0.9000
+# multid 0.2.0
 
-* Development version
 * Fixed bug: renaming output in D_regularized_fold functions
 * Fixed joining data frames by fold.var in D_regularized_fold functions
 * Added statistical inference to d_pooled_sd
 * Added probabilities of correct classification option to D_regularized_out and D_regularized_fold_out
 * Added area under the receiver operating characteristics to D_regularized_out and D_regularized_fold_out
+* Added probability classification tables to D_regularized_out and D_regularized_fold_out
+* Added more examples to README file
 
 # multid 0.1.0